requirements.txt
feedparser==6.0.11
pyshorteners==1.0.1
python-dotenv==1.0.1
tweepy==4.14.0
